检视问题修改 Signed-off-by: jw909 <909560957@qq.com>

+++ services/bundlemgr/src/app_control/app_jump_interceptor_event_subscriber.cpp
@@ -50,17 +50,21 @@ void AppJumpInterceptorEventSubscriber::OnReceiveEvent(const EventFwk::CommonEve
         return;
     }
     APP_LOGI("%{public}s, action:%{public}s.", __func__, action.c_str());
+    std::weak_ptr<AppJumpInterceptorEventSubscriber> weakThis = shared_from_this();
     if (action == EventFwk::CommonEventSupport::COMMON_EVENT_PACKAGE_REMOVED) {
         APP_LOGI("bundle remove, bundleName: %{public}s", bundleName.c_str());
-        auto task = [bundleName, db, userId]() {
+        auto task = [weakThis, bundleName, db, userId]() {
             BundleMemoryGuard memoryGuard;
             if (db == nullptr) {
                 APP_LOGE("Get invalid db");
                 return;
             }
-            APP_LOGI("start delete rule bundleName: %{public}s, userId:%d", bundleName.c_str(), userId);
-            db->DeleteRuleByCallerBundleName(bundleName, userId);
-            db->DeleteRuleByTargetBundleName(bundleName, userId);
+            std::shared_ptr<AppJumpInterceptorEventSubscriber> sharedThis = weakThis.lock();
+            if (sharedThis) {
+                APP_LOGI("start delete rule bundleName: %{public}s, userId:%d", bundleName.c_str(), userId);
+                db->DeleteRuleByCallerBundleName(bundleName, userId);
+                db->DeleteRuleByTargetBundleName(bundleName, userId);
+            }
         };
         eventHandler_->PostTask(task);
     } else {
